Immersive Weapons Fix.
Black sting is actually meant to be a one-handed sword with a dagger and a Greatsword version(named accordingly).

Also LWE/IW require a patch. Draedric Cresent is a One-hander in 'Immersive Weapons' but a greatsword in 'Lore Weapon Expansion',Unfortunately Greatsword bindings will always overwrite Arming Sword Bindings so either LWE needs a patch to add "Greatsword" or' IW needs a patch and there needs to be another binding section for binding overrides like this.